% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rimmer-package.R
\docType{package}
\name{rimmer-package}
\alias{rimmer}
\alias{rimmer-package}
\title{rimmer: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gorcha/rimmer}
  \item \url{https://gorcha.github.io/rimmer/}
  \item Report bugs at \url{https://github.com/gorcha/rimmer/issues}
}

}
\author{
\strong{Maintainer}: Danny Smith \email{danny@gorcha.org}

}
\keyword{internal}
